Allow local ref's to be served by nginx
This changeset uses the directory name of the SWAGGER_JSON file to use
as a root for all the files that have the following {json,yml,yaml}
extension.

By doing this one can use a docker-compose.yml like so:

```
environment:
  - "SWAGGER_JSON=/openapi/zd.yaml"
volumes:
  - "./openapi:/openapi"
```

or run docker like described in the installation.md file:
```
docker run -p 80:8080 \
  -e SWAGGER_JSON=/foo/swagger.json -v /bar:/foo swaggerapi/swagger-ui
```

Everything from /bar is mounted in /foo and /foo is than used by nginx
to serve it's openapi specification files.

Fixes: #5318, #5496, #4915

Signed-off-by: Wesley Schwengle <wesley@schwengle.net>